Slopes (#64)
* it's a start

* kind of works

* wip

* testing

* update x2 miscalculation

* moving up slopes almost works, one bug left to squash

* slopes ... seem to work ... one on side

* tile layout class improvement

* going up a slope on both left and right side works

* added break statements

* not sure...

* idk

* fix entire game breaking because I was off by one (lol)

* testing more slopes

* fixed jumping while going up slope

* move slope collision logic to map collision handler

* add slope y collision detection to x movement to make sure x movement follows slope bounds

* preliminary slopes work going downwards on left side, only x movement needed updates to achieve this

* slopes code massively refactored, fixed going down slopes to work when at top of slope (no longer temp go into air), made slopes work with variable layouts (stairs example)

* variable x movement works with slopes

* can go down slopes from right side

* variable slopes work, updated test map with new slopes

* slope glitch where player cannot stand on tip of a slope if there is not solid ground next to it has been fixed

* spacing

* variable slopes work and are no longer jittery -- fixed bug involving the scale slope array function

* update slope test map, comment up map collision handler better

* remove unused slope tiles

* cleaned up all slope test files, removed println statements used for debugging slopes, added comments

* remove unused code/comments/print statements from game object

+++ src/Level/Map.java
@@ -230,7 +230,7 @@ public abstract class Map {
     }
 
     // returns a tile based on a position in the map
-    public MapTile getTileByPosition(int xPosition, int yPosition) {
+    public MapTile getTileByPosition(float xPosition, float yPosition) {
         Point tileIndex = getTileIndexByPosition(xPosition, yPosition);
         if (isInBounds(Math.round(tileIndex.x), Math.round(tileIndex.y))) {
             return getMapTile(Math.round(tileIndex.x), Math.round(tileIndex.y));
